move prettyError into client
